requirements.txt
dnspython
netaddr
lxml
argparse
dnspython
slackclient
requests
colorama
flask
click
asyncio
uvloop
tqdm
aiodns
pyyaml
wrapt_timeout_decorator
needlepy
py-cord